refactor: test에 있던 파일들을 main으로 옮기고 디렉터리 구조 정리함, step1 리뷰 반영
