+++ src/test/kotlin/week1/step1/PersonTest.kt
@@ -1,4 +1,4 @@
-package study
+package week1.step1
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
