+++ src/test/kotlin/week1/step2/CalculatorTest.kt
@@ -1,4 +1,4 @@
-package week1
+package week1.step2
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
